+++ tests/e2e/cases/timers_promises_import/in.ts
@@ -6,12 +6,14 @@ let namedDelay = "";
 let namedImmediate = "";
 let namespaceDelay = "";
 let defaultImmediate = "";
+let optionDelay = "";
+let optionImmediate = "";
 
-delay(0, "named delay").then((value: string): void => {
+delay(0, "named delay", { signal: undefined, ref: false }).then((value: string): void => {
     namedDelay = value;
 });
 
-setImmediate("named immediate").then((value: string): void => {
+setImmediate("named immediate", { ref: true, signal: undefined }).then((value: string): void => {
     namedImmediate = value;
 });
 
@@ -23,7 +25,16 @@ timersPromises.setImmediate("default immediate").then((value: string): void => {
     defaultImmediate = value;
 });
 
+nodeTimersPromises.setTimeout(void 0, "option delay", { ref: undefined, signal: void 0 }).then((value: string): void => {
+    optionDelay = value;
+});
+
+timersPromises.setImmediate("option immediate", { ref: false }).then((value: string): void => {
+    optionImmediate = value;
+});
+
 console.log("named delay:", namedDelay);
 console.log("named immediate:", namedImmediate);
 console.log("namespace delay:", namespaceDelay);
 console.log("default immediate:", defaultImmediate);
+console.log("options:", optionDelay, optionImmediate);